+++ elements/x-swatch.js
@@ -26,6 +26,14 @@ export default class XSwatchElement extends HTMLElement {
       cursor: default;
       box-sizing: border-box;
       overflow: hidden;
+      /* Checkerboard pattern */
+      background-color: white;
+      background-size: 6px 6px;
+      background-position: 0 0, 0 3px, 3px -3px, -3px 0px;
+      background-image: linear-gradient(45deg, #d6d6d6 25%, transparent 25%),
+                        linear-gradient(-45deg, #d6d6d6 25%, transparent 25%),
+                        linear-gradient(45deg, transparent 75%, #d6d6d6 75%),
+                        linear-gradient(-45deg, transparent 75%, #d6d6d6 75%);
     }
 
     #preview {